move defaults tests
